quitando archivos no necesarios
